# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: MoxApp API
  description: |
    High-performance concurrent HTTP load testing tool with DNS timing metrics.
    
    This API provides endpoints for:
    - **Outgoing Traffic Management**: Configure and control external API calls
    - **Incoming Traffic Simulation**: Define mock routes with configurable response patterns
    - **Metrics & Monitoring**: Real-time metrics collection and health monitoring
    - **Scheduler Control**: Pause, resume, or emergency stop the load test
  version: 1.0.0
  contact:
    name: MoxApp Team
  license:
    name: Proprietary
    identifier: LicenseRef-MoxApp
servers:
  - url: __SERVER_URL__
    description: Current server
...
paths:
...
  /api/outgoing/endpoints:
    get:
      tags:
        - Outgoing Endpoints
      summary: List all outgoing endpoints
      description: Returns all configured outgoing endpoints
      operationId: listOutgoingEndpoints
      responses:
        '200':
          description: List of endpoints
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointListResponse'
    post:
      tags:
        - Outgoing Endpoints
      summary: Create outgoing endpoint
      description: Creates a new outgoing endpoint configuration
      operationId: createOutgoingEndpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndpointRequest'
      responses:
        '201':
          description: Endpoint created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Endpoint already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/endpoints/{name}:
    get:
      tags:
        - Outgoing Endpoints
      summary: Get outgoing endpoint by name
      description: Returns a specific outgoing endpoint configuration
      operationId: getOutgoingEndpoint
      parameters:
        - name: name
          in: path
          required: true
          description: Endpoint name
          schema:
            type: string
      responses:
        '200':
          description: Endpoint configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Endpoint'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Outgoing Endpoints
      summary: Update outgoing endpoint
      description: Updates an existing outgoing endpoint configuration
      operationId: updateOutgoingEndpoint
      parameters:
        - name: name
          in: path
          required: true
          description: Endpoint name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndpointRequest'
      responses:
        '200':
          description: Endpoint updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointUpdateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Outgoing Endpoints
      summary: Delete outgoing endpoint
      description: Deletes an outgoing endpoint by name
      operationId: deleteOutgoingEndpoint
      parameters:
        - name: name
          in: path
          required: true
          description: Endpoint name
          schema:
            type: string
      responses:
        '200':
          description: Endpoint deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/outgoing/control/endpoint:
    post:
      tags:
        - Outgoing Control
      summary: Enable/disable outgoing endpoint
      description: Enables or disables a specific outgoing endpoint
      operationId: setOutgoingEndpointEnabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - enabled
              properties:
                name:
                  type: string
                  description: Endpoint name
                enabled:
                  type: boolean
                  description: Enable or disable the endpoint
      responses:
        '200':
          description: Endpoint status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointEnableResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Endpoint:
      type: object
      properties:
        name:
          type: string
          description: Unique endpoint identifier
          example: user_api
        url_template:
          type: string
          description: URL template with placeholders
          example: "{{ env \"API_URL\" }}/users/{{ randomInt 1 1000 }}"
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH]
          example: GET
        frequency:
          type: number
          format: float
          description: Requests per minute
          example: 30
        auth:
          description: Authentication config reference or inline override
        timeout:
          type: integer
          description: Request timeout in seconds
          example: 30
        enabled:
          type: boolean
          description: Whether endpoint is active
          example: true
        body:
          type: string
          description: Request body template (for POST/PUT)
    EndpointRequest:
      type: object
      required:
        - name
        - url_template
        - method
        - frequency
      properties:
        name:
          type: string
          description: Unique endpoint identifier
          example: user_api
        url_template:
          type: string
          description: URL template with placeholders
          example: "{{ env \"API_URL\" }}/users/{{ randomInt 1 1000 }}"
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH]
          example: GET
        frequency:
          type: number
          format: float
          description: Requests per minute
          example: 30
        auth:
          description: Authentication config reference or inline override
        timeout:
          type: integer
          description: Request timeout in seconds
          example: 30
        enabled:
          type: boolean
          description: Whether endpoint is active
          default: true
        body:
          type: string
          description: Request body template
    EndpointListResponse:
      type: object
      properties:
        count:
          type: integer
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/Endpoint'
    EndpointCreateResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Endpoint created successfully
        endpoint:
          $ref: '#/components/schemas/Endpoint'
    EndpointUpdateResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Endpoint updated successfully
        endpoint:
          $ref: '#/components/schemas/Endpoint'
...
    EndpointEnableResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
        endpoint:
          type: string
        enabled:
          type: boolean
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
...